requirements.txt
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
mplfinance==0.12.10b0
plotly==5.22.0
streamlit==1.35.0
groq==0.11.0
python-dotenv==1.0.1
Pillow==10.3.0
requests==2.32.2
scikit-learn==1.4.2
xgboost==2.0.3